pyproject.toml
[tool.poetry]
name = "speccify"
version = "0.1.0"
description = ""
authors = ["Lyst Ltd. <devs@lyst.com>"]

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "*"
drf-yasg = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
pytest-django = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
